+++ pkg/testing/mock-supplychain.go
@@ -63,6 +63,50 @@ func completeLabels(workload v1alpha1.Workload, name string, kind string) map[st
 	return labels
 }
 
+func (i *MockSupplyChain) createTemplatingContext(workload v1alpha1.Workload, params map[string]apiextensionsv1.JSON) (map[string]interface{}, error) {
+	var (
+		inputs *Inputs
+		err    error
+	)
+
+	inputs = &Inputs{}
+
+	if i.BlueprintInputs != nil {
+		inputs, err = i.BlueprintInputs.GetBlueprintInputs()
+		if err != nil {
+			return nil, fmt.Errorf("get supply chain inputs: %w", err)
+		}
+	}
+
+	templatingContext := map[string]interface{}{
+		"workload": workload,
+		"params":   params,
+		"sources":  inputs.Sources,
+		"images":   inputs.Images,
+		"configs":  inputs.Configs,
+		//"deployment": // not implemented yet,
+	}
+
+	if len(inputs.Sources) == 1 {
+		for _, source := range inputs.Sources {
+			templatingContext["source"] = &source
+		}
+	}
+
+	if len(inputs.Images) == 1 {
+		for _, image := range inputs.Images {
+			templatingContext["image"] = image.Image
+		}
+	}
+
+	if len(inputs.Configs) == 1 {
+		for _, config := range inputs.Configs {
+			templatingContext["config"] = config.Config
+		}
+	}
+	return templatingContext, nil
+}
+
 type BlueprintParams interface {
 	GetBlueprintParams() ([]v1alpha1.BlueprintParam, error)
 }
@@ -173,48 +217,4 @@ func (p *BlueprintInputsFile) GetBlueprintInputs() (*Inputs, error) {
 	}
 
 	return &inputs, nil
-}
-
-func (i *MockSupplyChain) createTemplatingContext(workload v1alpha1.Workload, params map[string]apiextensionsv1.JSON) (map[string]interface{}, error) {
-	var (
-		inputs *Inputs
-		err    error
-	)
-
-	inputs = &Inputs{}
-
-	if i.BlueprintInputs != nil {
-		inputs, err = i.BlueprintInputs.GetBlueprintInputs()
-		if err != nil {
-			return nil, fmt.Errorf("get supply chain inputs: %w", err)
-		}
-	}
-
-	templatingContext := map[string]interface{}{
-		"workload": workload,
-		"params":   params,
-		"sources":  inputs.Sources,
-		"images":   inputs.Images,
-		"configs":  inputs.Configs,
-		//"deployment": // not implemented yet,
-	}
-
-	if len(inputs.Sources) == 1 {
-		for _, source := range inputs.Sources {
-			templatingContext["source"] = &source
-		}
-	}
-
-	if len(inputs.Images) == 1 {
-		for _, image := range inputs.Images {
-			templatingContext["image"] = image.Image
-		}
-	}
-
-	if len(inputs.Configs) == 1 {
-		for _, config := range inputs.Configs {
-			templatingContext["config"] = config.Config
-		}
-	}
-	return templatingContext, nil
 }
